Added prelim ARM template
Added the CreateSQL.json file that will create the SQL server (new) and the Return2Work.ps1 script to call it and also create the Resource Group for all the resources. Not prod ready because all assets are not yet created (nor the database) but adding these here for now.